Split service and controller into own tests
